ai/ml_api/requirements.txt
numpy
pandas
matplotlib
pythermalcomfort

pydantic
fastapi
uvicorn
torch

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
lightgbm>=4.0.0

joblib
seaborn
requests
python-dotenv
scikit-learn
